storage/bagger/src/requirements.txt
boto3
bagit
requests
fire
python-dateutil
